--- src/main/java/monsterboken/core/domain/Rarity.java
@@ -1,9 +1,0 @@
-package monsterboken.core.domain;
-
-public enum Rarity {
-    COMMON,
-    UNCOMMON,
-    RARE,
-    VERY_RARE,
-    UNIQUE;
-}
