+++ src/main/java/monsterboken/core/model/Mappable.java
@@ -1,4 +1,4 @@
-package monsterboken.core.domain;
+package monsterboken.core.model;
 
 public interface Mappable<T> {
 	T map();
