+++ src/main/java/monsterboken/core/model/Rarity.java
@@ -1,7 +1,7 @@
-package model;
+package monsterboken.core.model;
 
 public enum Rarity {
-	
+
 	COMMON, UNCOMMON, RARE, VERY_RARE, UNIQUE;
 
 }
